refactor: PathControllerMockTest 분리
